refactored to use context and set up demo functionality

+++ src/components/beer/beer.js
@@ -2,9 +2,11 @@ import React from 'react';
 import './beer.css';
 import config from '../../config';
 import TokenService from '../../services/token-service';
+import BeerContext from '../../BeerContext';
 
 class Beer extends React.Component {
-    
+    static contextType = BeerContext;
+
     state = {
         isExpanded: false,
     }
@@ -16,27 +18,33 @@ class Beer extends React.Component {
     }
 
     updateQuantity(newQuantity) {
-        this.setState({ quantity: newQuantity });
-        let updateFields = {
-            inventory_id: this.props.inventory,
-            updatedQuantity: newQuantity
-        };
-
-        fetch(`${config.API_ENDPOINT}/cellar/inventory`, {
-            method: 'PATCH',
-            headers: {
-                "content-type": "application/json",
-                "Authorization": `Bearer ${TokenService.getAuthToken()}`
-            },
-            body: JSON.stringify(updateFields)
-        })
-            .then(res => {
-                if(!res.ok) {
-                    throw new Error('Could not update quantity. Sorry bro...')
-                }
-                this.props.updateState(this.props.index, 'quantity', newQuantity)
+        if(TokenService.hasAuthToken()) {
+            //if logged in, update the quantity in the database
+            let updateFields = {
+                inventory_id: this.props.inventory,
+                updatedQuantity: newQuantity
+            };
+    
+            fetch(`${config.API_ENDPOINT}/cellar/inventory`, {
+                method: 'PATCH',
+                headers: {
+                    "content-type": "application/json",
+                    "Authorization": `Bearer ${TokenService.getAuthToken()}`
+                },
+                body: JSON.stringify(updateFields)
             })
-            .catch(err => console.log(err))
+                .then(res => {
+                    if(!res.ok) {
+                        throw new Error('Could not update quantity. Sorry bro...')
+                    }
+                    //if the API call is successful, also update in the app's state for quick rerender of the quantity amount
+                    this.context.updateSingleBeerInState(this.props.index, 'quantity', newQuantity)
+                })
+                .catch(err => console.log(err))
+        } else {
+            //if in demo mode, bypass the API call and just update the quantity in state
+            this.context.updateSingleBeerInState(this.props.index, 'quantity', newQuantity)
+        }
     }
 
     handleMinusClick = (e) => {
@@ -54,24 +62,30 @@ class Beer extends React.Component {
     }
 
     handleDelete = () => {
-        if(window.confirm('Are you sure you want to delete this beer?')) {
-            fetch(`${config.API_ENDPOINT}/cellar/inventory`, {
-                method: 'DELETE',
-                headers: {
-                    "content-type": "application/json",
-                    "Authorization": `Bearer ${TokenService.getAuthToken()}`
-                },
-                body: JSON.stringify({ inventory_id: this.props.inventory })
-            })
-                .then(res => {
-                    if(!res.ok) {
-                        throw new Error('Could not delete beer. Sorry bro...')
-                    }
-                    this.props.handleBeerDelete(this.props.inventory);
+        if(TokenService.hasAuthToken()) {
+            if(window.confirm('Are you sure you want to delete this beer?')) {
+                fetch(`${config.API_ENDPOINT}/cellar/inventory`, {
+                    method: 'DELETE',
+                    headers: {
+                        "content-type": "application/json",
+                        "Authorization": `Bearer ${TokenService.getAuthToken()}`
+                    },
+                    body: JSON.stringify({ inventory_id: this.props.inventory })
                 })
-                .catch(err => console.log(err))
+                    .then(res => {
+                        if(!res.ok) {
+                            throw new Error('Could not delete beer. Sorry bro...')
+                        }
+                        this.props.handleBeerDelete(this.props.inventory);
+                    })
+                    .catch(err => console.log(err))
+            } else {
+                console.log(`Good call, that's a good beer`);
+            }
         } else {
-            console.log(`Good call, that's a good beer`);
+            if(window.confirm('Are you sure you want to delete this beer?')) {
+                this.props.handleBeerDelete(this.props.inventory);
+            }
         }
     }
